CAL-116 Added MGMP Source Configuration, Integration Tests, Mock Server, and Updated Documentation
