+++ catalog/mgmp/catalog-source-mgmp/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,0 +1,81 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+/**
+ * Copyright (c) Codice Foundation
+ *
+ * This is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either
+ * version 3 of the License, or any later version. 
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details. A copy of the GNU Lesser General Public License is distributed along with this program and can be found at
+ * <http://www.gnu.org/licenses/lgpl.html>.
+ *
+ **/
+-->
+<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xmlns:cm="http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.1.0"
+           xsi:schemaLocation="http://www.osgi.org/xmlns/blueprint/v1.0.0
+           http://www.osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
+           http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.1.0
+           http://aries.apache.org/schemas/blueprint-cm/blueprint-cm-1.1.0.xsd">
+
+    <reference id="filterAdapter" interface="ddf.catalog.filter.FilterAdapter"/>
+    <reference id="filterBuilder" interface="ddf.catalog.filter.FilterBuilder"/>
+    <reference id="mimeTypeMapper" interface="ddf.mime.MimeTypeMapper"/>
+    <reference id="resourceReader" interface="ddf.catalog.resource.ResourceReader"
+               filter="(shortname=URLResourceReader)"/>
+    <reference id="cswTransformProvider" interface="com.thoughtworks.xstream.converters.Converter"/>
+    <reference id="securityManager" interface="ddf.security.service.SecurityManager"/>
+    <reference id="encryptionService" interface="ddf.security.encryption.EncryptionService"/>
+
+    <bean id="metacardTypes" class="ddf.catalog.util.impl.SortedServiceList"/>
+    <reference-list id="metacardTypeList" interface="ddf.catalog.data.MetacardType">
+        <reference-listener ref="metacardTypes" bind-method="bindPlugin"
+                            unbind-method="unbindPlugin"/>
+    </reference-list>
+
+    <cm:managed-service-factory
+            id="org.codice.alliance.catalog.mgmp"
+            factory-pid="Mgmp_Csw_Federated_Source"
+            interface="ddf.catalog.source.FederatedSource">
+        <cm:managed-component class="org.codice.ddf.spatial.ogc.csw.catalog.source.CswSourceImpl"
+                              init-method="init" destroy-method="destroy">
+            <property name="context" ref="blueprintBundleContext"/>
+            <property name="filterBuilder" ref="filterBuilder"/>
+            <property name="filterAdapter" ref="filterAdapter"/>
+            <property name="cswUrl" value=""/>
+            <property name="id" value=""/>
+            <property name="username" value=""/>
+            <property name="password" value=""/>
+            <property name="disableCnCheck" value="false"/>
+            <property name="coordinateOrder" value="LON_LAT"/>
+            <property name="usePosList" value="false"/>
+            <property name="metacardMappings">
+                <list/>
+            </property>
+            <property name="pollInterval" value="5"/>
+            <property name="resourceReader" ref="resourceReader"/>
+            <property name="outputSchema" value="http://www.isotc211.org/2005/gmd"/>
+            <property name="queryTypeName" value="gmd:MD_Metadata"/>
+            <property name="queryTypeNamespace" value="http://www.isotc211.org/2005/gmd"/>
+            <property name="cswTransformConverter">
+                <bean class="org.codice.ddf.spatial.ogc.csw.catalog.converter.GetRecordsResponseConverter">
+                    <argument ref="cswTransformProvider"/>
+                </bean>
+            </property>
+            <property name="isCqlForced" value="false"/>
+            <property name="forceSpatialFilter" value="NO_FILTER"/>
+            <property name="connectionTimeout" value="30000"/>
+            <property name="receiveTimeout" value="60000"/>
+            <property name="securityManager" ref="securityManager"/>
+            <property name="metacardTypes" ref="metacardTypes"/>
+            <property name="eventServiceAddress" value=""/>
+            <property name="registerForEvents" value="false"/>
+            <argument ref="encryptionService"/>
+            <cm:managed-properties persistent-id="" update-strategy="component-managed"
+                                   update-method="refresh"/>
+        </cm:managed-component>
+    </cm:managed-service-factory>
+
+</blueprint>
