+++ catalog/mgmp/catalog-source-mgmp/src/main/resources/OSGI-INF/metatype/metatype.xml
@@ -1,0 +1,105 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+/**
+ * Copyright (c) Codice Foundation
+ *
+ * This is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either
+ * version 3 of the License, or any later version. 
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details. A copy of the GNU Lesser General Public License is distributed along with this program and can be found at
+ * <http://www.gnu.org/licenses/lgpl.html>.
+ *
+ **/
+-->
+<metatype:MetaData xmlns:metatype="http://www.osgi.org/xmlns/metatype/v1.0.0">
+
+       <OCD name="MGMP CSW ISO Federated Source" id="Mgmp_Csw_Federated_Source"
+         description="CSW Federated Source using the MOD (MGMP) format">
+
+        <AD description="The unique name of the Source" name="Source ID" id="id" required="true"
+            type="String"/>
+
+        <AD description="URL to the endpoint implementing the Catalogue Service for Web (CSW) spec"
+            name="CSW URL" id="cswUrl" required="true" type="String"/>
+
+        <AD description="Username for CSW Service (optional)" name="Username" id="username"
+            required="false" type="String"/>
+        <AD description="Password for CSW Service (optional)" name="Password" id="password"
+            required="false" type="Password"/>
+
+        <AD description="Disable CN check for the server certificate. This should only be used when testing."
+            name="Disable CN Check" id="disableCnCheck" required="true"
+            type="Boolean" default="false"/>
+
+        <AD description="Coordinate order that remote source expects and returns spatial data in"
+            name="Coordinate Order" id="coordinateOrder" required="true"
+            type="String" default="LON_LAT">
+            <Option label="Lon/Lat" value="LON_LAT"/>
+            <Option label="Lat/Lon" value="LAT_LON"/>
+        </AD>
+
+        <AD description="Use a <posList> element rather than a series of <pos> elements when issuing geospatial queries containing a LinearRing"
+            name="Use posList in LinearRing" id="usePosList" required="false" type="Boolean"
+            default="false"/>
+
+        <AD description="Mapping of the Metacard Attribute names to their CSW property names. The format should be 'title=dc:title'."
+            name="Metacard Mappings" id="metacardMappings" required="false" type="String"
+            cardinality="100"
+            default="id=apiso:Identifier,effective=apiso:PublicationDate,created=apiso:CreationDate,modified=apiso:RevisionDate,title=apiso:AlternateTitle,AnyText=apiso:AnyText,ows:BoundingBox=apiso:BoundingBox">
+        </AD>
+
+        <AD description="Poll Interval to Check if the Source is available (in minutes - minimum 1)."
+            name="Poll Interval" id="pollInterval"
+            required="true" type="Integer" default="5"/>
+
+        <AD description="Amount of time to attempt to establish a connection before timing out, in milliseconds."
+            name="Connection Timeout" id="connectionTimeout"
+            required="true" type="Integer" default="30000"/>
+
+        <AD description="Amount of time to wait for a response before timing out, in milliseconds."
+            name="Receive Timeout" id="receiveTimeout"
+            required="true" type="Integer" default="60000"/>
+
+        <AD description="Output Schema" name="Output Schema" id="outputSchema" required="true"
+            type="String" default="http://www.isotc211.org/2005/gmd"/>
+
+        <AD description="Qualified Name for the Query Type used in the CSW GetRecords request"
+            name="Query Type Name" id="queryTypeName" required="true" type="String"
+            default="gmd:MD_Metadata"/>
+
+        <AD description="Namespace for the Query Type used in the CSW GetRecords request"
+            name="Query Type Namespace" id="queryTypeNamespace" required="true" type="String"
+            default="http://www.isotc211.org/2005/gmd"/>
+
+        <AD description="Force CQL Text" name="Force CQL Text as the Query Language"
+            id="isCqlForced" required="true"
+            type="Boolean" default="false"/>
+
+        <AD description="Force only the selected Spatial Filter Type as the only available Spatial Filter."
+            name="Forced Spatial Filter Type" id="forceSpatialFilter"
+            required="false" type="String" default="NO_FILTER">
+            <Option label="None" value="NO_FILTER"/>
+            <Option label="BBOX" value="BBOX"/>
+            <Option label="Beyond" value="Beyond"/>
+            <Option label="Contains" value="Contains"/>
+            <Option label="Crosses" value="Crosses"/>
+            <Option label="Disjoint" value="Disjoint"/>
+            <Option label="DWithin" value="DWithin"/>
+            <Option label="Intersects" value="Intersects"/>
+            <Option label="Equals" value="Equals"/>
+            <Option label="Overlaps" value="Overlaps"/>
+            <Option label="Touches" value="Touches"/>
+            <Option label="Within" value="Within"/>
+        </AD>
+
+        <AD description="Security attributes for this source" name="Security Attributes"
+            id="securityAttributeStrings" required="true" type="String" cardinality="1000"/>
+
+    </OCD>
+
+    <Designate pid="Mgmp_Csw_Federated_Source" factoryPid="Mgmp_Csw_Federated_Source">
+        <Object ocdref="Mgmp_Csw_Federated_Source"/>
+    </Designate>
+
+</metatype:MetaData>
