+++ catalog/mgmp/catalog-source-mgmp/src/main/resources/describable.properties
@@ -1,0 +1,4 @@
+version=${version}
+description=${description}
+organization=${organization.name}
+name=${artifactId}
