+++ catalog/mgmp/catalog-transformer-mgmp/src/test/resources/empty_mgmp.xml
