+++ catalog/mgmp/catalog-transformer-mgmp/src/test/resources/full_data_mgmp.xml
