+++ catalog/mgmp/catalog-transformer-mgmp/src/test/resources/sample_mgmp_1.xml
