+++ catalog/mgmp/catalog-transformer-mgmp/src/test/resources/sample_mgmp_2.xml
