+++ catalog/pom.xml
@@ -61,7 +61,7 @@
     <modules>
         <module>core</module>
         <module>nsili</module>
-        <module>transformer</module>
+        <module>mgmp</module>
         <module>imaging</module>
         <module>plugin</module>
         <module>video</module>
