+++ distribution/docs/src/main/resources/_contents/introduction-contents.adoc
@@ -9,3 +9,5 @@ Codice ${branding} is an open source, modular integration framework building on 
 ** Ingest and catalog NITF 2.1 imagery and build data-rich metacards for discovery and retrieval
 * STANAG 4609 KLV support
 ** Ingest FMV clips and extract KLV metadata for metacard creation
+* MGMP Ingest and Parsing
+** Ingest MGMP Records from a CSW Source and extract metadata for metacard creation
