+++ distribution/sdk/sample-mgmp-server/README.md
@@ -1,0 +1,126 @@
+#Mock MGMP Server
+
+##Mock MGMP Server
+
+The Mock MGMP server runs as a standalone HTTP server that accepts CSW and returns canned responses.  It is intended to provide a mock interface for MGMP. It was developed using the WireMock library.
+
+The Mock MGMP server uses JSON stubs which describe the mappings between requests and responses (see WireMock Stubbing for examples).  Below is the mapping for CSW GetCapabilities request to a CSW getCapabilities response (MgmpGetCapabilities.json).
+
+```json
+{
+    "request": {
+        "method": "GET",
+        "urlPath": "/csw",
+        "queryParameters": {
+          "acceptVersions": {
+            "matches": "2.0.2,2.0.1"
+          },
+          "request": {
+            "matches": "GetCapabilities"
+          },
+          "service": {
+            "matches": "CSW"
+          }
+        }
+    },
+    "response": {
+            "status": 200,
+            "headers" : {
+                "Content-Type": "text/xml"
+            },
+            "bodyFileName": "MgmpGetCapabilitiesResponse.xml"
+    }
+}
+```
+
+##Mock MGMP Server configuration
+
+The Mock MGMP Server configuration lives in the sample-mgmp-server pom.xml.  Things such as ports, etc. can be changed/added by modifying/adding arguments to the <arguments> section.  The server is currently set to run on port 9081.  See http://wiremock.org/running-standalone.html supported command line options.
+
+```xml
+    <plugin>
+        <groupId>org.codehaus.mojo</groupId>
+        <artifactId>exec-maven-plugin</artifactId>
+        <version>1.3.2</version>
+        <executions>
+            <execution>
+                <goals>
+                    <goal>java</goal>
+                </goals>
+            </execution>
+        </executions>
+        <configuration>
+           <includePluginDependencies>true</includePluginDependencies>
+           <includeProjectDependencies>false</includeProjectDependencies>
+           <executableDependency>
+               <groupId>com.github.tomakehurst</groupId>
+               <artifactId>wiremock</artifactId>
+           </executableDependency>
+           <mainClass>com.github.tomakehurst.wiremock.standalone.WireMockServerRunner
+           </mainClass>
+           <arguments>
+                <argument>--port</argument>
+                <argument>9081</argument>
+                <argument>--https-port</argument>
+                <argument>9443</argument>
+                <argument>--root-dir</argument>
+                <argument>src/main/resources/mock-mgmp</argument>
+                <argument>--verbose</argument>
+            </arguments>
+        </configuration>
+        <dependencies>
+            <dependency>
+                <groupId>com.github.tomakehurst</groupId>
+                <artifactId>wiremock</artifactId>
+                <classifier>standalone</classifier>
+                <version>1.53</version>
+            </dependency>
+        </dependencies>
+    </plugin>
+```
+
+##Using the Mock MGMP Server
+
+1. Navigate to the sample-mgmp-server directory.
+2. Run `mvn exec:java`
+3. Configure an MGMP Federated source using http://localhost:9081/csw and default settings
+
+Below is what one will see when the Mock MGMP Server is running.
+
+```
+$ mvn exec:java
+[INFO] Scanning for projects...
+[INFO]
+[INFO] ------------------------------------------------------------------------
+[INFO] Building Alliance :: Distribution :: SDK :: Sample MGMP Server 0.2-SNAPSHOT
+[INFO] ------------------------------------------------------------------------
+[INFO]
+[INFO] --- exec-maven-plugin:1.3.2:java (default-cli) @ sample-mgmp-server ---
+[WARNING] Warning: killAfter is now deprecated. Do you need it ? Please comment on MEXEC-6.
+2016-08-11 08:09:28.714 Verbose logging enabled
+2016-08-11 08:09:29.115:INFO::Logging to STDERR via org.mortbay.log.StdErrLog
+2016-08-11 08:09:29.120 Verbose logging enabled
+ /$$      /$$ /$$                     /$$      /$$                     /$$
+| $$  /$ | $$|__/                    | $$$    /$$$                    | $$
+| $$ /$$$| $$ /$$  /$$$$$$   /$$$$$$ | $$$$  /$$$$  /$$$$$$   /$$$$$$$| $$   /$$
+| $$/$$ $$ $$| $$ /$$__  $$ /$$__  $$| $$ $$/$$ $$ /$$__  $$ /$$_____/| $$  /$$/
+| $$$$_  $$$$| $$| $$  \__/| $$$$$$$$| $$  $$$| $$| $$  \ $$| $$      | $$$$$$/
+| $$$/ \  $$$| $$| $$      | $$_____/| $$\  $ | $$| $$  | $$| $$      | $$_  $$
+| $$/   \  $$| $$| $$      |  $$$$$$$| $$ \/  | $$|  $$$$$$/|  $$$$$$$| $$ \  $$
+|__/     \__/|__/|__/       \_______/|__/     |__/ \______/  \_______/|__/  \__/
+
+port:                         9081
+https-port:                   9443
+https-keystore:               (null)
+enable-browser-proxying:      false
+no-request-journal:           false
+verbose:                      true
+```
+
+##Shutting Down the Mock Server
+
+One can shutdown the Mock Server by sending a POST request with an empty body to `http://localhost:9081/__admin/shutdown`.  NOTE: Tools such as Postman or cURL can be used to POST the request.
+
+```
+$ curl -X POST http://localhost:9081/__admin/shutdown
+```
